extract helper class and neighborListOf method
